Simplify administration
This patch is intended to simplfy the content administration. The about and
contact file paths were hard-coded and placed withing the gallery directory.
The pattern for ignored files was changed to leading underscore from dotfile so
that the files are visible even on *nix systems. The cover image and logo
image were moved out of the skin directory and the default skin will now assume
that there is a file named _cover.jpg within the gallery directory (it's up to
the user to put it there). Logo has been completely omitted as there is no way
to easily change it by the content administrator.

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>