Add static site generator
This commit adds a static site generator for seagull galleries. The generator
is activated using the --generate-site or -G switch. When activated, the app
will dump static versions of the main gallery pages into the gallery directory
and will not start the server. The pages are named 'index.html', 'page2.html',
and so on.

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>