Add --templates option
This commit adds the --templates options which prints out the absolute paths to
template directories in use by the app. In order to implement this, a new
appconf module has been introduced: appconf.commands, and the command line
arguments from the seagull.main are relayed through the app object to this
module.

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>