Implement watch/stop/recompile commands in seagull
Instead of using the setup module for these commands, this commit moves it to
the Seagull command framework. Three new seagull utility commands are
introduced that match the previous setup commands:

- watch
- stop-watchers
- recompile

The watch and recompile commands support a ``--skin-path`` command which is a
more reliable way to specify the override for the skin path to use. The default
is to use the skin specified by the configuration file, and this option forces
the use of an alternative path.

Signed-off-by: Hajime Yamasaki Vukelic <hayavuk@gmail.com>